python-service/requirements.txt
# Core dependencies
torch>=2.0.0
numpy>=1.24.0

# SMPL model
smplx>=0.1.28

# 3D mesh processing
trimesh>=4.0.0
pygltflib>=1.16.0

# Measurement and anthropometry
scikit-learn>=1.3.0
scipy>=1.11.0

# Web service
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# Utilities
python-multipart>=0.0.6
tqdm>=4.66.0
Pillow>=10.0.0
